backend/requirements.txt
Flask
pydantic
Flask-Cors
Flask-SQLAlchemy
Flask-Login
Flask-SocketIO==5.3.6
python-socketio[client]==5.11.0
python-engineio==4.9.0
python-dotenv
google-cloud-bigquery
google-auth-oauthlib
google-api-core
requests
APScheduler
PyMySQL
SQLAlchemy
sqlglot
boto3
google-cloud-storage
google-cloud-pubsub
google-cloud-eventarc
azure-storage-blob
azure-storage-file-share
azure-data-tables
azure-storage-queue
pyarrow
pandas
openpyxl
fastavro
msgpack
pymongo
deltalake
pyiceberg
pyyaml
tomli
h5py
netcdf4
tensorflow
pyreadstat

